ENH: If input image type not supported, try to default to float component
Not all image type are wrapped in Python for ITK. When the image type
is not found, try to default to `float` as component type. If the
conversion is not possible, the script will still fail, but in
most cases it will work (e.g. If the input image component type is
`int`).